requirements.txt
email-validator>=2.2.0
flask-migrate>=4.1.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-jwt-extended>=4.7.1
flask-marshmallow>=1.3.0
flask-restx>=1.3.0
flask-cors>=5.0.1
marshmallow-sqlalchemy>=1.4.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.1.0
sqlalchemy>=2.0.40
werkzeug>=3.1.3
marshmallow-enum>=1.5.1
requests>=2.32.3
